+++ weblate/formats/__init__.py
@@ -16,7 +16,3 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-"""File format specific behavior."""
-
-
-default_app_config = "weblate.formats.apps.FormatsConfig"