+++ weblate/lang/management/commands/move_language.py
@@ -56,7 +56,7 @@ class Command(BaseCommand):
 
         for plural in source.plural_set.iterator():
             try:
-                new_plural = target.plural_set.get(formula=plural.formula)
+                new_plural = target.plural_set.filter(formula=plural.formula).first()
                 plural.translation_set.update(plural=new_plural)
             except Plural.DoesNotExist:
                 plural.language = target