+++ weblate/trans/models/component.py
@@ -62,6 +62,7 @@ from weblate.trans.models.translation import Translation
 from weblate.trans.models.variant import Variant
 from weblate.trans.signals import (
     component_post_update,
+    store_post_load,
     translation_post_add,
     vcs_post_commit,
     vcs_post_push,
@@ -1144,6 +1145,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
         """
 
         def add(repo):
+            self.log_info("checking for key to add for %s", repo)
             parsed = urlparse(repo)
             if not parsed.hostname:
                 parsed = urlparse(f"ssh://{repo}")
@@ -1152,6 +1154,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
                     port = parsed.port
                 except ValueError:
                     port = ""
+                self.log_info("adding SSH key for %s:%s", parsed.hostname, port)
                 add_host_key(None, parsed.hostname, port)
 
         add(self.repo)
@@ -1390,8 +1393,8 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
                 self.delete_alert("PushFailure")
                 return True
             except RepositoryException as error:
-                report_error(cause="Could not push the repo")
                 error_text = self.error_text(error)
+                report_error(cause="Could not push the repo")
                 Change.objects.create(
                     action=Change.ACTION_FAILED_PUSH,
                     component=self,
@@ -2035,7 +2038,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
         if was_change:
             self.update_variants()
             component_post_update.send(sender=self.__class__, component=self)
-            self.sync_terminology()
+            self.schedule_sync_terminology()
 
         self.unload_sources()
         self.run_batched_checks()
@@ -2566,7 +2569,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
             self.create_glossary()
 
             # Make sure all languages are present
-            self.sync_terminology()
+            self.schedule_sync_terminology()
 
             # Run automatically installed addons. They are run upon installation,
             # but there are no translations created at that point.
@@ -2957,7 +2960,16 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
             messages.error(request, _("Translation file already exists!"))
         else:
             with self.repository.lock:
-                file_format.add_language(fullname, language, base_filename)
+                file_format.add_language(
+                    fullname,
+                    language,
+                    base_filename,
+                    callback=lambda store: store_post_load.send(
+                        sender=translation.__class__,
+                        translation=translation,
+                        store=store,
+                    ),
+                )
                 if send_signal:
                     translation_post_add.send(
                         sender=self.__class__, translation=translation
@@ -3037,7 +3049,7 @@ class Component(FastDeleteModelMixin, models.Model, URLMixin, PathMixin, CacheKe
                 result[installed.event].append(addon)
         return result
 
-    def sync_terminology(self):
+    def schedule_sync_terminology(self):
         """Trigger terminology sync in the background."""
         from weblate.glossary.tasks import sync_terminology