+++ weblate/vcs/ssh.py
@@ -24,6 +24,7 @@ import subprocess
 from base64 import b64decode, b64encode
 from distutils.spawn import find_executable
 
+from django.core.management.utils import find_command
 from django.utils.functional import cached_property
 from django.utils.translation import gettext as _
 
@@ -192,44 +193,63 @@ def can_generate_key():
     return find_executable("ssh-keygen") is not None
 
 
+SSH_WRAPPER_TEMPLATE = r"""#!/bin/sh
+exec {command} \
+    -o "UserKnownHostsFile={known_hosts}" \
+    -o "IdentityFile={identity}" \
+    -o StrictHostKeyChecking=yes \
+    -o HashKnownHosts=no \
+    -o UpdateHostKeys=yes \
+    -F /dev/null \
+    "$@"
+"""
+
+
 class SSHWrapper:
     # Custom ssh wrapper
     # - use custom location for known hosts and key
     # - do not hash it
     # - strict hosk key checking
     # - force not using system configuration (to avoid evil things as SendEnv)
-    SSH_WRAPPER_TEMPLATE = r"""#!/bin/sh
-    exec ssh \
-        -o "UserKnownHostsFile={known_hosts}" \
-        -o "IdentityFile={identity}" \
-        -o StrictHostKeyChecking=yes \
-        -o HashKnownHosts=no \
-        -F /dev/null \
-        "$@"
-    """
 
     @cached_property
-    def filename(self):
-        """Calculates unique wrapper filename.
+    def digest(self):
+        return calculate_checksum(SSH_WRAPPER_TEMPLATE, data_dir("ssh"))
+
+    @property
+    def path(self):
+        """Calculates unique wrapper path.
 
         It is based on template and DATA_DIR settings.
         """
-        digest = calculate_checksum(self.SSH_WRAPPER_TEMPLATE, data_dir("ssh"))
-        return ssh_file(f"ssh-weblate-wrapper-{digest}")
+        return ssh_file(f"bin-{self.digest}")
+
+    @property
+    def filename(self):
+        """Calculates unique wrapper filename."""
+        return os.path.join(self.path, "ssh")
 
     def create(self):
         """Create wrapper for SSH to pass custom known hosts and key."""
-        if os.path.exists(self.filename):
-            return
+        if not os.path.exists(self.path):
+            os.makedirs(self.path)
 
-        with open(self.filename, "w") as handle:
-            handle.write(
-                self.SSH_WRAPPER_TEMPLATE.format(
-                    known_hosts=ssh_file(KNOWN_HOSTS), identity=ssh_file(RSA_KEY)
+        for command in ("ssh", "scp"):
+            filename = os.path.join(self.path, command)
+
+            if os.path.exists(filename):
+                continue
+
+            with open(filename, "w") as handle:
+                handle.write(
+                    SSH_WRAPPER_TEMPLATE.format(
+                        command=find_command(command),
+                        known_hosts=ssh_file(KNOWN_HOSTS),
+                        identity=ssh_file(RSA_KEY),
+                    )
                 )
-            )
 
-        os.chmod(self.filename, 0o755)  # nosec
+            os.chmod(filename, 0o755)  # nosec
 
 
 SSH_WRAPPER = SSHWrapper()