+++ weblate/wladmin/tests.py
@@ -21,6 +21,7 @@ import json
 import os
 
 import responses
+import social_django.utils
 from django.conf import settings
 from django.core import mail
 from django.core.checks import Critical
@@ -37,6 +38,8 @@ from weblate.utils.checks import check_data_writable
 from weblate.utils.unittest import tempdir_setting
 from weblate.wladmin.middleware import ManageMiddleware
 from weblate.wladmin.models import BackupService, ConfigurationError, SupportStatus
+
+TEST_BACKENDS = ("weblate.accounts.auth.WeblateUserBackend",)
 
 
 class AdminTest(ViewTestCase):
@@ -198,11 +201,51 @@ class AdminTest(ViewTestCase):
                 "email": "noreply@example.com",
                 "username": "username",
                 "full_name": "name",
+                "send_email": 1,
             },
             follow=True,
         )
-        self.assertContains(response, "User has been invited")
+        self.assertContains(response, "Created user account")
         self.assertEqual(len(mail.outbox), 1)
+
+    def test_invite_user_nosend(self):
+        response = self.client.get(reverse("manage-users"))
+        self.assertContains(response, "E-mail")
+        response = self.client.post(
+            reverse("manage-users"),
+            {
+                "email": "noreply@example.com",
+                "username": "username",
+                "full_name": "name",
+            },
+            follow=True,
+        )
+        self.assertContains(response, "Created user account")
+        self.assertEqual(len(mail.outbox), 0)
+
+    @override_settings(AUTHENTICATION_BACKENDS=TEST_BACKENDS)
+    def test_invite_user_nomail(self):
+        try:
+            # psa creates copy of settings...
+            orig_backends = social_django.utils.BACKENDS
+            social_django.utils.BACKENDS = TEST_BACKENDS
+
+            response = self.client.get(reverse("manage-users"))
+            self.assertContains(response, "E-mail")
+            response = self.client.post(
+                reverse("manage-users"),
+                {
+                    "email": "noreply@example.com",
+                    "username": "username",
+                    "full_name": "name",
+                    "send_email": 1,
+                },
+                follow=True,
+            )
+            self.assertContains(response, "Created user account")
+            self.assertEqual(len(mail.outbox), 1)
+        finally:
+            social_django.utils.BACKENDS = orig_backends
 
     def test_check_user(self):
         response = self.client.get(