+++ weblate/wladmin/views.py
@@ -29,6 +29,8 @@ from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.utils import timezone
 from django.utils.decorators import method_decorator
+from django.utils.html import escape
+from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
 from django.utils.translation import gettext_lazy
 from django.views.decorators.http import require_POST
@@ -344,8 +346,17 @@ class AdminUserList(UserList):
         if "email" in request.POST:
             invite_form = AdminInviteUserForm(request.POST)
             if invite_form.is_valid():
-                invite_form.save(request)
-                messages.success(request, _("User has been invited to this project."))
+                user = invite_form.save(request)
+                messages.success(
+                    request,
+                    mark_safe(
+                        escape(_("Created user account %s."))
+                        % '<a href="{}">{}</a>'.format(
+                            escape(user.get_absolute_url()),
+                            escape(user.username),
+                        )
+                    ),
+                )
                 return redirect("manage-users")
         return super().post(request, **kwargs)