+++ docs/changes.rst
@@ -10,6 +10,7 @@ Not yet released.
 * Improved SAML authentication integration.
 * Fixed :ref:`vcs-gerrit` integration to better handle corner cases.
 * Weblate now requires Django 3.2.
+* Fixed inviting users when e-mail authentication is disabled.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/64?closed=1>`__.