+++ docs/changes.rst
@@ -11,6 +11,7 @@ Not yet released.
 * Fixed :ref:`vcs-gerrit` integration to better handle corner cases.
 * Weblate now requires Django 3.2.
 * Fixed inviting users when e-mail authentication is disabled.
+* Improved language definitions.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/64?closed=1>`__.